requirements-cu118.txt
--index-url https://download.pytorch.org/whl/cu118
--extra-index-url https://pypi.org/simple

torch==2.5.1+cu118
torchvision==0.20.1+cu118

ultralytics>=8.3.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
lap>=0.5.12
